Added auth. py and main .py
auth.py
This class allows you to authrorize into the google api client.It
only needs to be done once.For the purpose of this appliation,
scopes need to be provided. This is done when the instance is
initalised.

main.py
This module allows to create folders,media files and list folders in
you google drive.